Integrating TCP transport
Finished refactoring, integrated TCP transport and updated test suite.
