+++ src/stun_client.js
@@ -1,147 +1,127 @@
 'use strict'
 
-var events = require('events')
-var util = require('util')
+var inherits = require('util').inherits
 var Q = require('q')
 var winston = require('winston')
 
+var Attributes = require('./attributes')
 var Packet = require('./packet')
-var UdpTransport = require('./transports/udp')
+var StunComm = require('./stun_comm')
 
-// Init client object
+// Constructor
 var StunClient = function (stunHost, stunPort, transport) {
-  if (stunPort === undefined || stunHost === undefined) {
-    var error = '[stun-js] stun host and/or port are undefined'
-    winston.error(error)
-    throw new Error(error)
-  }
-  this._responseCallbacks = {}
-
-  events.EventEmitter.call(this)
-
-  this._transport = new UdpTransport(stunHost, stunPort)
-  this._transport.onMessage(this.onIncomingMessage())
-  this._transport.onError(this.onFailure())
-  this._transport.init()
+  StunComm.call(this, stunHost, stunPort, transport)
 }
 
-// Inherit EventEmitter
-util.inherits(StunClient, events.EventEmitter)
+// Inherit from StunComm
+inherits(StunClient, StunComm)
 
-// Close client
-StunClient.prototype.close = function () {
-  winston.debug('[stun-js] closing client')
-  this._transport.release()
-}
+/** StunComm operations */
 
-/** UDP communication */
-
-// Send STUN request
-StunClient.prototype.sendStunRequest = function (bytes, onResponse, onFailure) {
-  // get tid
-  var tid = bytes.readUInt32BE(16)
-  // store response handler for this tid
-  this._responseCallbacks[tid] = onResponse
-  // send bytes
-  this._transport.send(
-    bytes,
-    function () { // on success
-      // do nothing
-    },
-    function (error) { // on failure
-      onFailure(error)
-    }
-  )
-}
-
-StunClient.prototype.sendStunRequestP = function (bytes) {
-  var deferred = Q.defer()
-  this.sendStunRequest(
-    bytes,
-    function(stunResponse) { // on response
-      deferred.resolve(stunResponse)
-    },
-    function(error) { // on failure
-      deferred.reject(error)
-    }
-  )
-  return deferred.promise
-}
-
-// Send STUN indication
-StunClient.prototype.sendStunIndicationP = function (bytes) {
-  return this._transport.sendP(bytes)
-}
-
-StunClient.prototype.sendStunIndication = function (bytes, onSuccess, onFailure) {
-  if (onSuccess === undefined || onFailure === undefined) {
-    var error = '[stun-js] send stun indication callback handlers are undefined'
-    winston.error(error)
-    throw new Error(error)
-  }
-  this._transport.send(bytes, onSuccess, onFailure)
-}
-
-// Incoming message handler
-StunClient.prototype.onIncomingMessage = function () {
-  var self = this
-  return function (bytes, rinfo) {
-    winston.debug('[stun-js] receiving message from ' + JSON.stringify(rinfo))
-    // this is a stun packet
-    var stunPacket = Packet.decode(bytes)
-    if (stunPacket) {
-      switch (stunPacket.type) {
-        case Packet.TYPE.SUCCESS_RESPONSE:
-          self.onIncomingStunResponse(stunPacket, rinfo)
-          break
-        case Packet.TYPE.ERROR_RESPONSE:
-          self.onIncomingStunResponse(stunPacket, rinfo)
-          break
-        case Packet.TYPE.INDICATION:
-          self.onIncomingStunIndication(stunPacket, rinfo)
-          break
-        default:
-          var errorMsg = "[stun-js] don't know how to process incoming STUN message -- dropping it on the floor"
-          winston.error(errorMsg)
-          throw new Error(errorMsg)
+// Bind request
+StunClient.prototype.bindP = function () {
+  return this.sendBindRequestP()
+    .then(function (bindReply) {
+      var errorCode = bindReply.getAttribute(Attributes.ERROR_CODE)
+      // check if the reply includes an error code attr
+      if (errorCode) {
+        throw new Error('[stun-js] bind error: ' + errorCode.reason)
       }
-    } else {
-      self.onOtherIncomingMessage(bytes, rinfo)
-    }
+      var mappedAddressAttr = bindReply.getAttribute(Attributes.XOR_MAPPED_ADDRESS)
+      if (!mappedAddressAttr) {
+        mappedAddressAttr = bindReply.getAttribute(Attributes.MAPPED_ADDRESS)
+      }
+      var mappedAddress = {
+        address: mappedAddressAttr.address,
+        port: mappedAddressAttr.port
+      }
+      return Q.fcall(function () {
+        return mappedAddress
+      })
+    })
+}
+
+StunClient.prototype.bind = function (onSuccess, onFailure) {
+  if (onSuccess === undefined || onFailure === undefined) {
+    var error = '[stun-js] bind callback handlers are undefined'
+    winston.error(error)
+    throw new Error(error)
   }
+  this.bindP()
+    .then(function (result) {
+      onSuccess(result)
+    })
+    .catch(function (error) {
+      onFailure(error)
+    })
 }
 
-// Incoming STUN reply
-StunClient.prototype.onIncomingStunResponse = function (stunPacket, rinfo) {
-  // this is a stun reply
-  var onResponseCallback = this._responseCallbacks[stunPacket.tid]
-  if (onResponseCallback) {
-    onResponseCallback(stunPacket)
-    delete this._responseCallbacks[stunPacket.tid]
-  } else {
-    var errorMsg = '[stun-js] no handler available to process response with tid ' + stunPacket.tid
-    winston.error(errorMsg)
-    throw new Error(errorMsg)
+/** Message transmission */
+
+// Send STUN bind request
+StunClient.prototype.sendBindRequestP = function () {
+  winston.debug('[stun-js] send bind request (using promises)')
+  var message = composeBindRequest()
+  return this.sendStunRequestP(message)
+}
+
+StunClient.prototype.sendBindRequest = function (onSuccess, onFailure) {
+  winston.debug('[stun-js] send bind request')
+  if (onSuccess === undefined || onFailure === undefined) {
+    var error = '[stun-js] send bind request callback handlers are undefined'
+    winston.error(error)
+    throw new Error(error)
   }
+  this.sendBindRequestP()
+    .then(function (reply) {
+      onSuccess(reply)
+    })
+    .catch(function (error) {
+      onFailure(error)
+    })
 }
 
-// Incoming STUN indication
-StunClient.prototype.onIncomingStunIndication = function (stunPacket, rinfo) {
-  this.emit('indication', stunPacket, rinfo)
+// Send STUN bind indication
+StunClient.prototype.sendBindIndicationP = function () {
+  winston.debug('[stun-js] send bind indication (using promises)')
+  var message = composeBindIndication()
+  return this.sendStunIndicationP(message)
 }
 
-// Incoming message that is different from regular STUN packets
-StunClient.prototype.onOtherIncomingMessage = function (bytes, rinfo) {
-  this.emit('message', bytes, rinfo)
-}
-
-// Error handler
-StunClient.prototype.onFailure = function () {
-  return function (error) {
-    var errorMsg = '[stun-js] client error: ' + error
-    winston.error(errorMsg)
-    throw new Error(errorMsg)
+StunClient.prototype.sendBindIndication = function (onSuccess, onFailure) {
+  winston.debug('[stun-js] send bind indication')
+  if (onSuccess === undefined || onFailure === undefined) {
+    var error = '[stun-js] send bind indication callback handlers are undefined'
+    winston.error(error)
+    throw new Error(error)
   }
+  this.sendBindIndicationP()
+    .then(function (reply) {
+      onSuccess(reply)
+    })
+    .catch(function (error) {
+      onFailure(error)
+    })
+}
+
+/** Message composition */
+
+// Create bind request message
+function composeBindRequest () {
+  // create packet
+  var packet = new Packet(Packet.METHOD.BINDING, Packet.TYPE.REQUEST)
+  // encode packet
+  var message = packet.encode()
+  return message
+}
+
+// Create bind request message
+function composeBindIndication () {
+  // create packet
+  var packet = new Packet(Packet.METHOD.BINDING, Packet.TYPE.INDICATION)
+  // encode packet
+  var message = packet.encode()
+  return message
 }
 
 module.exports = StunClient
